Updated Process for Docstamp and Updated the DocStamp Crystal report

+++ ProducersBank/Models/PriceListModel.cs
@@ -11,7 +11,7 @@ namespace ProducersBank.Models
         public string Bank { get; set; }
         public string ChequeDescription { get; set; }
         public string ChkType { get; set; }
-        public  double DocStampPrice { get; set; }
+        public  int DocStampPrice { get; set; }
         public double unitprice{ get; set; }
 
     }
